backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic-settings
python-dotenv
python-multipart
sqlalchemy
chromadb
sentence-transformers
pymupdf
groq
langchain
streamlit